# src/api/openapi.yaml
openapi: 3.0.3
info:
  title: Stock Scanner API
  description: |
    Real-time stock market analysis API providing market movers, value analysis, options scanning,
    historical data, sector trends, and macro indicators.

    **Caching Behavior:**
    - Market movers: 5 minutes
    - Value analysis: 30 minutes
    - Options data: 1 minute (real-time sensitive)
    - History, sectors, macro: No cache
  version: 1.0.0
  contact:
    name: Stock Scanner API
servers:
  - url: /
    description: Current origin (recommended)
  - url: http://localhost:3000
    description: Local development server
...
paths:
  /api/movers:
    get:
      tags:
        - Market Data
      summary: Get market movers
      description: |
        Fetches top gaining, losing, or most active stocks with enriched value metrics and money flow data.
        Results are cached for 5 minutes to reduce API load.
      operationId: getMarketMovers
      parameters:
        - name: type
          in: query
          description: Type of market movers to fetch
          required: false
          schema:
            type: string
            enum: [gainers, losers, active]
            default: active
        - name: limit
          in: query
          description: Number of stocks to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Successful response with enriched market movers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EnrichedMover'
              example:
                - symbol: AAPL
                  name: Apple Inc.
                  price: 185.50
                  change: 5.25
                  changesPercentage: 2.91
                  volume: 58234567
                  valueScore: 75.3
                  valueMetrics:
                    peRatio: 29.5
                    pbRatio: 45.2
                    debtToEquity: 1.73
                    roe: 0.456
                    currentRatio: 0.98
                  moneyFlowStrength: 68.5
                  sector: Technology
                  industry: Consumer Electronics
                  reasons:
                    - Strong revenue growth
                    - High profitability
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/value/{symbol}:
    get:
      tags:
        - Analysis
      summary: Analyze stock value
      description: |
        Performs comprehensive value analysis on a stock symbol including fundamentals,
        ratios, and scoring. Results are cached for 30 minutes.
      operationId: getValueAnalysis
      parameters:
        - name: symbol
          in: path
          description: Stock ticker symbol (uppercase)
          required: true
          schema:
            type: string
            pattern: '^[A-Z]{1,5}$'
            example: AAPL
      responses:
        '200':
          description: Value analysis result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValueScore'
              example:
                symbol: AAPL
                score: 75.3
                sector: Technology
                industry: Consumer Electronics
                metrics:
                  peRatio: 29.5
                  pbRatio: 45.2
                  debtToEquity: 1.73
                  roe: 0.456
                  currentRatio: 0.98
                reasons:
                  - Strong revenue growth
                  - High profitability
                  - Excellent return on equity
        '404':
          description: Stock data not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Analysis failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/options/{symbol}:
    get:
      tags:
        - Analysis
      summary: Scan options activity
      description: |
        Scans options flow for a symbol, detecting unusual activity, sentiment, and signals.
        Results are cached for 1 minute due to real-time sensitivity. Automatically saves
        results to historical database.
      operationId: scanOptions
      parameters:
        - name: symbol
          in: path
          description: Stock ticker symbol (uppercase)
          required: true
          schema:
            type: string
            pattern: '^[A-Z]{1,5}$'
            example: AAPL
      responses:
        '200':
          description: Options scan result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OptionsScanResult'
              example:
                moneyFlowStrength: 68.5
                signals:
                  - symbol: AAPL
                    type: call
                    strike: 190
                    expiryISO: '2024-03-15T00:00:00Z'
                    volume: 5234
                    openInterest: 12456
                    last: 8.50
                    bid: 8.45
                    ask: 8.55
                    mid: 8.50
                    notional: 44489000
                    pos: 0.85
                    direction: buy
                    lastTradeISO: '2024-03-01T15:30:00Z'
                    ageMin: 5
                    rmp: 185.50
                    moneyness: 1.024
                    daysToExpiry: 14
                    marketCap: 2850000000000
                    notionalToMarketCap: 0.0000156
                    isDeepOTMPut: false
                    tenorBucket: short
                    isShortTermSpec: true
                    isLongTermHedge: false
                    comboId: ''
                    comboType: standalone
                    isComboHedge: false
                    isLargeOTMCall: false
                    hedgeScore: 0.12
                    hedgeTags: []
                    iv: 0.28
                    spreadPct: 0.59
                    traderType: institutional
                    spotConfirmation: strong
                    daysToEarnings: null
                    directionConfidence: 0.89
                    signalQuality: 0.85
                sentiment:
                  bullishNotional: 145000000
                  bearishNotional: 78000000
                  putCallRatio: 0.54
                  sentiment: 0.67
                marketState: regular
                rmp: 185.50
        '500':
          description: Options scan failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/history/{symbol}:
    get:
      tags:
        - History
      summary: Get historical scan results
      description: |
        Retrieves historical options scan results for a symbol. No caching applied.
      operationId: getHistory
      parameters:
        - name: symbol
          in: path
          description: Stock ticker symbol (uppercase)
          required: true
          schema:
            type: string
            pattern: '^[A-Z]{1,5}$'
            example: AAPL
      responses:
        '200':
          description: Historical scan records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HistoryRecord'
              example:
                - timestamp: '2024-03-01T15:30:00Z'
                  symbol: AAPL
                  sentiment:
                    bullishNotional: 145000000
                    bearishNotional: 78000000
                    putCallRatio: 0.54
                    sentiment: 0.67
                  signalCount: 12
                  topSignals:
                    - type: call
                      strike: 190
                      notional: 44489000
                      direction: buy
        '500':
          description: Failed to fetch history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/trends/sectors:
    get:
      tags:
        - Analytics
      summary: Get sector trends
      description: |
        Analyzes sector performance trends over the past 14 days. No caching applied.
      operationId: getSectorTrends
      responses:
        '200':
          description: Sector trend analysis
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SectorTrend'
              example:
                - sector: Technology
                  avgChange: 2.45
                  stockCount: 87
                  momentum: strong
                  topGainers:
                    - NVDA
                    - AAPL
                    - MSFT
                - sector: Healthcare
                  avgChange: -0.52
                  stockCount: 65
                  momentum: weak
                  topGainers:
                    - JNJ
                    - PFE
        '500':
          description: Failed to fetch sector trends
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/macro:
    get:
      tags:
        - Analytics
      summary: Get macro snapshot
      description: |
        Retrieves current macro market indicators including indices, DXY, VIX, and regime.
      operationId: getMacroSnapshot
      responses:
        '200':
          description: Macro market snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MacroSnapshot'
              example:
                indices:
                  - symbol: ^GSPC
                    name: S&P 500
                    price: 5234.18
                    change: 45.23
                    changePercent: 0.87
                  - symbol: ^DJI
                    name: Dow Jones
                    price: 38456.78
                    change: 125.67
                    changePercent: 0.33
                  - symbol: ^IXIC
                    name: NASDAQ
                    price: 16234.56
                    change: 98.45
                    changePercent: 0.61
                dxy:
                  price: 103.45
                  change: 0.12
                  changePercent: 0.12
                vix:
                  price: 14.56
                  change: -0.45
                  changePercent: -3.00
                overallRegime: bullish
        '500':
          description: Failed to fetch macro snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/news:
    get:
      tags:
        - News
      summary: Get FinancialJuice news
      description: |
        Returns latest FinancialJuice news by category with optional cache bypass.
        When request fails but cache exists, stale cache may be returned.
      operationId: getFinancialJuiceNews
      parameters:
        - name: category
          in: query
          description: News category
          required: false
          schema:
            type: string
            enum: [macro, indices, equities, crypto]
            default: macro
        - name: limit
          in: query
          description: Maximum items to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 20
        - name: fresh
          in: query
          description: Force fresh fetch and bypass in-memory cache
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: News list response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsListResponse'
              example:
                category: macro
                cached: false
                count: 2
                items:
                  - id: '921381'
                    timestamp: '2026-02-09T13:05:12.000Z'
                    headline: 'Fed official says inflation progress remains uneven'
                    category: 'Economy'
                    urgency: 3
                    source: 'FinancialJuice'
                  - id: '921382'
                    timestamp: '2026-02-09T13:10:45.000Z'
                    headline: 'US treasury yields rise ahead of auction'
                    category: 'Bonds'
                    urgency: 3
                    source: 'FinancialJuice'
        '500':
          description: Failed to load news
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/strategies/credit-spreads:
    get:
      tags:
        - Strategies
      summary: Get ranked directional credit spread candidates
      description: |
        Builds a directional credit spread snapshot by combining market movers,
        daily chart structure, macro regime, and near-term vertical spread templates.
        The schema is written for machine readers: downstream AI clients should use
        `strategyType`, `direction`, `anchorType`, `anchorLevel`, `invalidationPrice`,
        and `blockers` to reason about the setup instead of inferring intent from UI copy.
      operationId: getCreditSpreadStrategySnapshot
      parameters:
        - in: query
          name: strategyType
          required: false
          description: |
            Strategy playbook to rank. `BEAR_CALL_CREDIT` targets bearish resistance-fade setups.
            `BULL_PUT_CREDIT` targets bullish support-hold setups.
          schema:
            type: string
            enum: [BEAR_CALL_CREDIT, BULL_PUT_CREDIT]
            default: BEAR_CALL_CREDIT
        - in: query
          name: setupState
          required: false
          description: |
            Optional server-side execution filter. `ACTIONABLE` means the setup passed
            structure, macro, and spread-template gates. `WATCHLIST` means the setup is
            informative but currently blocked.
          schema:
            type: string
            enum: [ACTIONABLE, WATCHLIST]
        - in: query
          name: limit
          required: false
          description: Maximum number of candidates returned after filtering.
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Ranked credit spread strategy snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditSpreadStrategySnapshot'
              example:
                generatedAt: '2026-03-12T08:10:00.000Z'
                strategyType: BEAR_CALL_CREDIT
                macro:
                  indices: []
                  dxy:
                    symbol: DX-Y.NYB
                    price: 104.0
                    changePercent: 0.5
                    trend: UP
                  vix:
                    symbol: ^VIX
                    price: 20.0
                    changePercent: 6.0
                    status: RISING
                  overallRegime: RISK_OFF
                filters:
                  minPrice: 20
                  maxPrice: 500
                  minVolume: 20000000
                  targetDteMin: 3
                  targetDteMax: 7
                candidates:
                  - strategyType: BEAR_CALL_CREDIT
                    direction: BEARISH
                    symbol: TSLA
                    name: Tesla
                    price: 212
                    changePercent: -8.1
                    volume: 120000000
                    score: 35
                    setupState: ACTIONABLE
                    structureScore: 28
                    macroScore: 6
                    valueBias: 2
                    anchorType: RESISTANCE
                    anchorLevel: 226
                    invalidationPrice: 228.26
                    volumeRatio20: 2.31
                    closeLocationValue: 0.09
                    upperWickRatio: 0.09
                    lowerWickRatio: 0.09
                    eventTags: []
                    thesis:
                      - Closed below the prior 20-session low
                      - Options chain offers a liquid call credit spread above resistance
                    blockers: []
                    watchlistReasons: []
                    dte: 5
                    spreadTemplate:
                      strategyType: BEAR_CALL_CREDIT
                      shortLegType: CALL
                      longLegType: CALL
                      expiryISO: '2026-03-20'
                      shortStrike: 230
                      longStrike: 235
                      width: 5
                      dte: 5
                      shortDelta: 0.16
                      shortBid: 1.9
                      shortAsk: 2.2
                      longBid: 0.45
                      longAsk: 0.65
                      shortMid: 2.05
                      longMid: 0.55
                      creditMid: 1.5
                      creditPctWidth: 0.3
                      takeProfitAt: 0.6
                      stopLossAt: 3.0
        '500':
          description: Failed to build credit spread strategy snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/strategies/call-credit:
    get:
      tags:
        - Strategies
      summary: Get ranked call credit spread candidates
      deprecated: true
      description: |
        Legacy alias for `/api/strategies/credit-spreads?strategyType=BEAR_CALL_CREDIT`.
        New AI and frontend consumers should prefer the unified `credit-spreads` route.
      operationId: getLegacyCallCreditStrategySnapshot
      responses:
        '200':
          description: Ranked bearish call credit strategy snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditSpreadStrategySnapshot'
        '500':
          description: Failed to build call credit strategy snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    EnrichedMover:
      type: object
      description: Market mover enriched with value metrics and money flow data
      properties:
        symbol:
          type: string
          description: Stock ticker symbol
          example: AAPL
        name:
          type: string
          description: Company name
          example: Apple Inc.
        price:
          type: number
          format: float
          description: Current stock price
          example: 185.50
        change:
          type: number
          format: float
          description: Price change in dollars
          example: 5.25
        changesPercentage:
          type: number
          format: float
          description: Price change percentage
          example: 2.91
        volume:
          type: integer
          description: Trading volume
          example: 58234567
        valueScore:
          type: number
          format: float
          nullable: true
          description: Value score (0-100), null if unavailable
          example: 75.3
        valueMetrics:
          $ref: '#/components/schemas/ValueMetrics'
          nullable: true
        moneyFlowStrength:
          type: number
          format: float
          description: Money flow indicator strength (0-100)
          example: 68.5
        sector:
          type: string
          description: Company sector
          example: Technology
        industry:
          type: string
          description: Company industry
          example: Consumer Electronics
        reasons:
          type: array
          items:
            type: string
          description: Reasons for value score
          example:
            - Strong revenue growth
            - High profitability
...